# docs/user_openapi.yaml
openapi: 3.0.3
info:
  title: user API
  description: Auto-generated API documentation for user
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Development server

paths:
  /api/v1/user:
    get:
      summary: RegisterRoutes
      description: RegisterRoutes
      tags:
        - user
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example: {
                "success": true,
                "data": {
                  "id": 1,
                  "name": "Example User",
                  "email": "user@example.com",
                  "age": 25,
                  "created_at": "2023-10-20T10:00:00Z",
                  "updated_at": "2023-10-20T10:00:00Z"
                }
              }
        '201':
          description: Created
        '400':
          description: Bad Request
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
  /api/v1/users:
    get:
      summary: Get all users
      description: Get list of users with pagination and filtering
      tags:
        - user
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example: {
                "success": true,
                "data": {
                  "id": 1,
                  "name": "Example User",
                  "email": "user@example.com",
                  "age": 25,
                  "created_at": "2023-10-20T10:00:00Z",
                  "updated_at": "2023-10-20T10:00:00Z"
                }
              }
        '201':
          description: Created
        '400':
          description: Bad Request
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
    post:
      summary: Create a new user
      description: Create a new user with the provided data
      tags:
        - user
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example: {
                "success": true,
                "data": {
                  "id": 1,
                  "name": "Example User",
                  "email": "user@example.com",
                  "age": 25,
                  "created_at": "2023-10-20T10:00:00Z",
                  "updated_at": "2023-10-20T10:00:00Z"
                }
              }
        '201':
          description: Created
        '400':
          description: Bad Request
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
  /api/v1/users/{id}:
    get:
      summary: Get user by ID
      description: Get user details by ID
      tags:
        - user
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example: {
                "success": true,
                "data": {
                  "id": 1,
                  "name": "Example User",
                  "email": "user@example.com",
                  "age": 25,
                  "created_at": "2023-10-20T10:00:00Z",
                  "updated_at": "2023-10-20T10:00:00Z"
                }
              }
        '201':
          description: Created
        '400':
          description: Bad Request
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
    put:
      summary: Update user
      description: Update user details
      tags:
        - user
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example: {
                "success": true,
                "data": {
                  "id": 1,
                  "name": "Example User",
                  "email": "user@example.com",
                  "age": 25,
                  "created_at": "2023-10-20T10:00:00Z",
                  "updated_at": "2023-10-20T10:00:00Z"
                }
              }
        '201':
          description: Created
        '400':
          description: Bad Request
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
    delete:
      summary: Delete user
      description: Delete user by ID
      tags:
        - user
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example: {
                "success": true,
                "data": {
                  "id": 1,
                  "name": "Example User",
                  "email": "user@example.com",
                  "age": 25,
                  "created_at": "2023-10-20T10:00:00Z",
                  "updated_at": "2023-10-20T10:00:00Z"
                }
              }
        '201':
          description: Created
        '400':
          description: Bad Request
        '404':
          description: Not Found
        '500':
          description: Internal Server Error

components:
  schemas:
    Response:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
        error:
          type: string
        meta:
          type: object
          properties:
            page:
              type: integer
            page_size:
              type: integer
            total:
              type: integer
            pages:
              type: integer
    CreateUserRequest:
      type: object
      properties:
        name:
          type: string
          required: true
          example: John Doe
        email:
          type: string
          required: true
          example: user@example.com
        age:
          type: integer
          required: true
          example: 123
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
          required: true
          example: John Doe
        email:
          type: string
          required: true
          example: user@example.com
        age:
          type: integer
          required: true
          example: 123
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          required: true
          example: 1
        name:
          type: string
          required: true
          example: John Doe
        email:
          type: string
          required: true
          example: user@example.com
        age:
          type: integer
          required: true
          example: 123
        created_at:
          type: string
          required: true
          example: 2023-10-20T10:00:00Z
        updated_at:
          type: string
          required: true
          example: 2023-10-20T10:00:00Z
    UserHandler:
      type: object
      properties:
        service:
          type: string
          example: 
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: John Doe
        email:
          type: string
          example: user@example.com
        age:
          type: integer
          example: 123
        created_at:
          type: string
          example: 2023-10-20T10:00:00Z
        updated_at:
          type: string
          example: 2023-10-20T10:00:00Z
        -:
          type: string
          example: example